added some tests for mongoRoute
