+++ test/models/book.spec.js
